+++ tests/integration/dbcrash/db.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Wikimedia\Rdbms;
+
+interface IReadableDatabase {
+	public function select(
+		$table, $vars, $conds = '', $fname = __METHOD__,
+		$options = [], $join_conds = []
+	);
+}
+
+interface IDatabase extends  IReadableDatabase {
+}