+++ tests/integration/getqueryinfocrash/db.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Wikimedia\Rdbms;
+
+interface IDatabase {
+	public function select(
+		$table, $vars, $conds = '', $fname = __METHOD__,
+		$options = [], $join_conds = []
+	);
+}